+++ mods/src/text-input-box/TextInputScreen.cpp
@@ -1,11 +1,9 @@
-#include <libreborn/libreborn.h>
-
 #include <mods/text-input-box/TextInputScreen.h>
 
 // VTable
 void TextInputScreen::setup(Screen_vtable *vtable) {
     static Screen_keyPressed_t original_keyPressed = vtable->keyPressed;
-    vtable->keyPressed = [](Screen *super2, int key) {
+    vtable->keyPressed = [](Screen *super2, const int key) {
         original_keyPressed(super2, key);
         TextInputScreen *self = (TextInputScreen *) super2;
         for (int i = 0; i < int(self->m_textInputs->size()); i++) {
@@ -14,7 +12,7 @@ void TextInputScreen::setup(Screen_vtable *vtable) {
         }
     };
     static Screen_keyboardNewChar_t original_keyboardNewChar = vtable->keyboardNewChar;
-    vtable->keyboardNewChar = [](Screen *super2, char key) {
+    vtable->keyboardNewChar = [](Screen *super2, const char key) {
         original_keyboardNewChar(super2, key);
         TextInputScreen *self = (TextInputScreen *) super2;
         for (int i = 0; i < int(self->m_textInputs->size()); i++) {
@@ -23,7 +21,7 @@ void TextInputScreen::setup(Screen_vtable *vtable) {
         }
     };
     static Screen_mouseClicked_t original_mouseClicked = vtable->mouseClicked;
-    vtable->mouseClicked = [](Screen *super2, int x, int y, int param_1) {
+    vtable->mouseClicked = [](Screen *super2, const int x, const int y, const int param_1) {
         original_mouseClicked(super2, x, y, param_1);
         TextInputScreen *self = (TextInputScreen *) super2;
         for (int i = 0; i < int(self->m_textInputs->size()); i++) {
@@ -32,7 +30,7 @@ void TextInputScreen::setup(Screen_vtable *vtable) {
         }
     };
     static Screen_render_t original_render = vtable->render;
-    vtable->render = [](Screen *super2, int x, int y, float param_1) {
+    vtable->render = [](Screen *super2, const int x, const int y, const float param_1) {
         original_render(super2, x, y, param_1);
         TextInputScreen *self = (TextInputScreen *) super2;
         for (int i = 0; i < int(self->m_textInputs->size()); i++) {